refactor(STT): make icon red when recording. consolidate logic to AudioRecorder component
